requirements.txt
openai
streamlit
pymupdf
pillow
pandas